# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "iPF-Fund-request-real",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "local Server"
        },
        {
            "url": "http://127.0.0.1:8000/",
            "description": "local Server"
        }
    ],
    "paths": {
        "/api/Admin/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "user login",
                "operationId": "1babe736c625a7665caf168fd8443246",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "email",
                                        "format": "string",
                                        "default": "example@gmail.com"
                                    },
                                    "password": {
                                        "description": "password",
                                        "format": "string",
                                        "default": "xx"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "logged in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "token",
                                            "format": "string",
                                            "default": "komcqwmcfoqwmfcfoqwkcmwocmqoefmweokmf"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "wrong crediantials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Invalid email or password"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Admin/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "user register",
                "operationId": "3e746c359ed9f6e6085f5e3a50d54999",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "description": "full_name",
                                        "format": "string",
                                        "default": "example"
                                    },
                                    "email": {
                                        "description": "email",
                                        "format": "string",
                                        "default": "example@gmail.com"
                                    },
                                    "phone_no": {
                                        "description": "phone_no",
                                        "format": "string",
                                        "default": "07xxxxxxxx"
                                    },
                                    "password": {
                                        "description": "password",
                                        "format": "string",
                                        "default": "****"
                                    },
                                    "role": {
                                        "description": "role",
                                        "format": "string",
                                        "default": "staff"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "registered successfully"
                                        },
                                        "token": {
                                            "description": "token",
                                            "format": "string",
                                            "default": "ksnfclKFMCLkmcklaCMAKLSCMKLMCKLASCM"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Fields required",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/Admin/invitation": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "user invitation",
                "operationId": "571e9ee7c86fb07a2b467b947882cc03",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "description": "full_name",
                                        "format": "string",
                                        "default": "example"
                                    },
                                    "email": {
                                        "description": "email",
                                        "format": "string",
                                        "default": "example@gmail.com"
                                    },
                                    "phone_no": {
                                        "description": "phone_no",
                                        "format": "string",
                                        "default": "07xxxxxxxx"
                                    },
                                    "role": {
                                        "description": "role",
                                        "format": "string",
                                        "default": "staff"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "invitated successfully"
                                        },
                                        "users": {
                                            "properties": {
                                                "full_name": {
                                                    "description": "full_name",
                                                    "format": "string",
                                                    "default": "example"
                                                },
                                                "email": {
                                                    "description": "email",
                                                    "format": "string",
                                                    "default": "example@gmail.com"
                                                },
                                                "phone_no": {
                                                    "description": "phone_no",
                                                    "format": "string",
                                                    "default": "07xxxxxxxx"
                                                },
                                                "role": {
                                                    "description": "role",
                                                    "format": "string",
                                                    "default": "staff"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Fields required",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}